Moves unit tests of task requests to a single file (#1669)
* refactor : moves unit tests of task requests to one file

* chore : adds space between describe and it blocks in task requests
